+++ common/src/main/java/io/github/mortuusars/horseman/world/LeavesCollisionMode.java
@@ -1,0 +1,7 @@
+package io.github.mortuusars.horseman.world;
+
+public enum LeavesCollisionMode {
+    FULL,
+    IGNORE_LOWEST_BLOCK,
+    IGNORE_ALL;
+}